# The ruv-swarm system now has enterprise-grade database connection pooling that eliminates the connection exhaustion and deadlock issues identified in the original production readiness assessment. The system is fully ready for high-concurrency production workloads.