… ruv-swarm/npm/data/ruv-swarm.db-shm
… ruv-swarm/npm/docs/CONNECTION-POOL.md
… ruv-swarm/npm/src/persistence-pooled.js
… ruv-swarm/npm/src/sqlite-pool.js
… ruv-swarm/npm/test/sqlite-pool-ha.test.js